Moved config handling to TraitToEmbedTransformer
